refs #26353. Restructured file
